+++ packages/sit-onyx/src/components/OnyxButton/OnyxButton.vue
@@ -17,7 +17,9 @@ const props = withDefaults(defineProps<OnyxButtonProps>(), {
 const { densityClass } = useDensity(props);
 
 const emit = defineEmits<{
-  /** Emitted when the button is clicked (and is not disabled). */
+  /**
+   * Emitted when the button is clicked (and is not disabled).
+   */
   click: [];
 }>();
 </script>
@@ -41,7 +43,7 @@ const emit = defineEmits<{
   >
     <OnyxIcon v-if="props.icon && !props.loading" :icon="props.icon" />
     <OnyxLoadingIndicator v-if="props.loading" class="onyx-button__loading" />
-    <span v-else class="onyx-button__label onyx-truncation-ellipsis">{{ props.label }}</span>
+    <span class="onyx-button__label onyx-truncation-ellipsis">{{ props.label }}</span>
   </button>
 </template>
 
@@ -75,6 +77,7 @@ const emit = defineEmits<{
     --onyx-button-text-color: var(--onyx-color-text-icons-primary-intense);
     --onyx-button-outline-color: var(--onyx-color-base-primary-200);
 
+    position: relative;
     display: inline-flex;
     height: var(--onyx-button-height);
     max-width: 100%;
@@ -195,9 +198,12 @@ const emit = defineEmits<{
       line-height: 1.5rem;
     }
 
+    &--loading &__label {
+      visibility: hidden;
+    }
+
     &__loading {
-      width: var(--onyx-spacing-3xl);
-      height: 100%;
+      position: absolute;
     }
   }