+++ packages/sit-onyx/src/components/OnyxTextarea/OnyxTextarea.stories.ts
@@ -116,7 +116,7 @@ export const HiddenLabel = {
  * This example shows a textarea with a custom error message.
  * Will only be shown after user interaction.
  */
-export const CustomError = {
+export const CustomError: Story = {
   args: {
     ...Default.args,
     customError: {
@@ -125,7 +125,13 @@ export const CustomError = {
     },
     placeholder: "Interact with me to show error",
   },
-} satisfies Story;
+  decorators: [
+    (story) => ({
+      components: { story },
+      template: `<div style="padding-bottom: 2rem"> <story /> </div>`,
+    }),
+  ],
+};
 
 /**
  * This example shows a textarea with custom autosize settings (min=2 rows, max=12 rows).