+++ packages/sit-onyx/src/components/OnyxTimer/OnyxTimer.vue
@@ -15,7 +15,7 @@ const emit = defineEmits<{
 }>();
 
 const { locale } = injectI18n();
-const { timeLeft, isEnded } = useTimer({ endTime: computed(() => props.endTime) });
+const { timeLeft, isEnded } = useTimer(computed(() => props.endTime));
 
 const timeFormat = computed(
   () => new Intl.RelativeTimeFormat(locale.value, { numeric: "always", style: "short" }),
@@ -55,6 +55,8 @@ watchEffect(() => isEnded.value && emit("timerEnded"));
     padding: var(--onyx-spacing-2xs) var(--onyx-spacing-md);
     font-family: var(--onyx-font-family);
     color: var(--onyx-color-text-icons-neutral-intense);
+    width: max-content;
+    max-width: 100%;
 
     &__label {
       color: var(--onyx-color-text-icons-neutral-soft);