+++ packages/sit-onyx/src/components/OnyxToast/OnyxToast.vue
@@ -8,7 +8,6 @@ import { computed } from "vue";
 import { useDensity } from "../../composables/density";
 import { injectI18n } from "../../i18n";
 import OnyxIcon from "../OnyxIcon/OnyxIcon.vue";
-import OnyxToastProgressBar from "../OnyxToastProgressBar/OnyxToastProgressBar.vue";
 import type { OnyxToastProps } from "./types";
 
 const props = withDefaults(defineProps<OnyxToastProps>(), {
@@ -52,7 +51,7 @@ const icon = computed(() => {
     :is="props.clickable ? 'button' : 'div'"
     class="onyx-toast"
     :class="[`onyx-toast--${props.color}`, densityClass]"
-    role="alert"
+    :role="props.color === 'danger' || props.color === 'warning' ? 'alert' : 'status'"
     :aria-label="props.clickable ? props.headline : undefined"
     @click="props.clickable && emit('click')"
   >
@@ -83,12 +82,15 @@ const icon = computed(() => {
       </div>
     </div>
 
-    <OnyxToastProgressBar
+    <!-- key is used to restart the animation when the duration changes -->
+    <time
       v-if="hasProgressBar"
-      :color="props.color"
-      :duration="props.duration"
-      @timer-ended="emit('close')"
-    />
+      :key="props.duration"
+      aria-hidden="true"
+      class="onyx-toast__progress-bar"
+      :style="{ animationDuration: `${props.duration}ms` }"
+      @animationend="emit('close')"
+    ></time>
   </component>
 </template>
 
@@ -137,10 +139,12 @@ const icon = computed(() => {
     position: relative;
     z-index: var(--onyx-z-index-notification);
 
-    .onyx-toast-progress-bar {
-      position: absolute;
-      bottom: 0;
-      left: 0;
+    &:hover,
+    &:focus,
+    &:has(&__description:focus-visible) {
+      .onyx-toast__progress-bar {
+        animation-play-state: paused;
+      }
     }
 
     &:is(button) {
@@ -149,9 +153,13 @@ const icon = computed(() => {
       border: none;
       padding: 0;
 
-      &:focus-visible {
+      &:focus-within {
         outline: 0.25rem solid var(--onyx-toast-outline-color);
       }
+    }
+
+    &:has(&__description:focus-visible) {
+      outline: 0.25rem solid var(--onyx-toast-outline-color);
     }
 
     &__wrapper {
@@ -177,6 +185,27 @@ const icon = computed(() => {
       }
     }
 
+    &__progress-bar {
+      display: block;
+      height: var(--onyx-spacing-5xs);
+      background-color: var(--onyx-toast-progress-bar-color);
+      animation: onyx-toast-progress-bar linear;
+      width: 0;
+
+      position: absolute;
+      bottom: 0;
+      left: 0;
+
+      @keyframes onyx-toast-progress-bar {
+        0% {
+          width: 100%;
+        }
+        100% {
+          width: 0%;
+        }
+      }
+    }
+
     &__close {
       background: none;
       padding: 0;
@@ -198,7 +227,7 @@ const icon = computed(() => {
     }
 
     // only show close button on hover if it closes automatically
-    &:has(.onyx-toast-progress-bar) {
+    &:has(.onyx-toast__progress-bar) {
       &:not(:hover) .onyx-toast__close {
         display: none;
       }