+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,36 @@
 # alpha-test-app
 
+## 0.1.0-alpha.147
+
+### Patch Changes
+
+- Updated dependencies [0bdb49a]
+- Updated dependencies [0bdb49a]
+  - sit-onyx@1.0.0-alpha.150
+
+## 0.1.0-alpha.146
+
+### Patch Changes
+
+- Updated dependencies [1cc020a]
+- Updated dependencies [1cc020a]
+  - sit-onyx@1.0.0-alpha.149
+
+## 0.1.0-alpha.145
+
+### Patch Changes
+
+- Updated dependencies [2d0458d]
+  - sit-onyx@1.0.0-alpha.148
+
+## 0.1.0-alpha.144
+
+### Patch Changes
+
+- Updated dependencies [743ee88]
+- Updated dependencies [b7e9aaf]
+  - sit-onyx@1.0.0-alpha.147
+
 ## 0.1.0-alpha.143
 
 ### Patch Changes